Continuous fill scale that aligns with the Urban Institute style

Continuous fill scale that aligns with the Urban Institute style

scale_fill_gradientn(..., colours = c("#CFE8F3", "#A2D4EC", "#73BFE2",
  "#46ABDB", "#1696D2", "#12719E", "#0A4C6A", "#062635"),
  colors = c("#CFE8F3", "#A2D4EC", "#73BFE2", "#46ABDB", "#1696D2",
  "#12719E", "#0A4C6A", "#062635"), values = NULL, space = "Lab",
  na.value = "grey50", guide = "colourbar")

Arguments

...
other arguments passed to discrete_scale()

colours
vector of colours

colors
vector of colours

values
if colours should not be evenly positioned along the gradient this vector gives the position (between 0 and 1) for each colour in the colours vector. See rescale for a convience function to map an arbitrary range to between 0 and 1

space
colour space in which to calculate gradient. Must be "Lab" - other values are deprecated.

na.value
default color for NA values

guide
legend representation of scale